cf_launch_test/requirements.txt
functions-framework==3.2.1
flask==2.3.2
google-cloud-error-reporting==1.6.0
google-cloud-pubsub